% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_content_urls.R
\name{get_all_content_urls}
\alias{get_all_content_urls}
\title{get_all_content_urls}
\usage{
get_all_content_urls(journal_home_page, article_type)
}
\arguments{
\item{journal_home_page}{journal_home_page}

\item{article_type}{article type}
}
\value{
a vector of all cotnent pages of a journal's specific article type
}
\description{
get_all_content_urls
}
\examples{
get_all_content_urls('Cell Research','review')
}
