% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_paper_metrics.R
\name{get_all_paper_metrics}
\alias{get_all_paper_metrics}
\title{get_all_paper_metrics}
\usage{
get_all_paper_metrics(paper_urls, sleep_seconds = 10)
}
\arguments{
\item{paper_urls}{All paper urls}

\item{sleep_seconds}{sleep_seconds}
}
\value{
Return a dataframe of all papers' metrics
}
\description{
get_all_paper_metrics
}
\examples{
get_all_paper_metrics(c('https://www.nature.com/articles/s41421-018-0016-3',
'https://www.nature.com/articles/s41421-022-00458-3'))
}
