% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_paper_urls.R
\name{get_all_paper_urls}
\alias{get_all_paper_urls}
\title{get_all_paper_urls}
\usage{
get_all_paper_urls(content_urls, sleep_seconds = 10)
}
\arguments{
\item{content_urls}{Many content urls}

\item{sleep_seconds}{sleep_seconds}
}
\value{
Return paper urls of all content urls
}
\description{
get_all_paper_urls
}
\examples{
all_content_urls <- get_nature_journal_all_content_urls('STTT','comment')
get_all_paper_urls(all_content_urls)
}
