% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_Phenomics_paper_metrics.R
\name{get_all_Phenomics_paper_metrics}
\alias{get_all_Phenomics_paper_metrics}
\title{get_all_Phenomics_paper_metrics}
\usage{
get_all_Phenomics_paper_metrics(online_data, sleep_seconds = 10)
}
\arguments{
\item{online_data}{get on web}

\item{sleep_seconds}{sleep_seconds}
}
\value{
Return a dataframe of all Phenomics papers' metrics
}
\description{
get_all_Phenomics_paper_metrics
}
\examples{
todo
}
