% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journal_homepage.R
\name{get_journal_homepage}
\alias{get_journal_homepage}
\title{Get journal homepage}
\usage{
get_journal_homepage(journal_name)
}
\arguments{
\item{journal}{journal name}
}
\value{
Return the journal homepage
}
\description{
Get journal homepage
}
\examples{
get_journal_homepage('Phenomics')
}
