% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nature_journal_all_content_urls.R
\name{get_nature_journal_all_content_urls}
\alias{get_nature_journal_all_content_urls}
\title{get nature journal all content urls}
\usage{
get_nature_journal_all_content_urls(journal_home_page, article_type)
}
\arguments{
\item{journal_home_page}{journal_home_page}

\item{article_type}{article type}
}
\value{
a vector of all cotnent pages of a journal's specific article type
}
\description{
get nature journal all content urls
}
\examples{
get_nature_journal_all_content_urls('Cell Research','review')
}
