% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nature_journal_all_content_pages.R
\name{get_nature_journal_all_cotnent_pages}
\alias{get_nature_journal_all_cotnent_pages}
\title{get nature journal all cotnent pages}
\usage{
get_nature_journal_all_cotnent_pages(journal_name, article_type)
}
\arguments{
\item{journal_name}{journal name}

\item{article_type}{article type}
}
\value{
a vector of all cotnent pages of a journal's specific article type
}
\description{
get nature journal all cotnent pages
}
\examples{
get_nature_journal_all_cotnent_pages('Cell Research','review')
}
