% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_paper_metrics.R
\name{get_paper_metrics}
\alias{get_paper_metrics}
\title{get_paper_metrics}
\usage{
get_paper_metrics(paper_url)
}
\arguments{
\item{paper_url}{paper_url}
}
\value{
Return paper metrics
}
\description{
get_paper_metrics
}
\examples{
get_paper_metrics('https://www.nature.com/articles/s41421-018-0016-3')
}
