% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_paper_urls.R
\name{get_paper_urls}
\alias{get_paper_urls}
\title{get_paper_urls}
\usage{
get_paper_urls(content_url)
}
\arguments{
\item{content_url}{one content_url}
}
\value{
Return the paper urls on content_url
}
\description{
get_paper_urls
}
\examples{
get_paper_urls('https://www.nature.com/celldisc/news-and-comment?searchType=journalSearch&sort=PubDate&page=8')
}
