% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_Phenomics_papers_metrics.R
\name{get_Phenomics_paper_metrics}
\alias{get_Phenomics_paper_metrics}
\title{get_Phenomics_paper_metrics}
\usage{
get_Phenomics_paper_metrics(online_data, idx)
}
\arguments{
\item{online_data}{online_data from}

\item{idx}{index}
}
\value{
Return paper metrics
}
\description{
get_Phenomics_paper_metrics
}
\examples{
todo
}
